requirements/requirements.txt
numpy
Pillow
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
factory-boy==2.12.0
Faker==4.0.1
gunicorn==20.0.4
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
django-minio-storage
#django-extensions
django-taggit
django-taggit-autosuggest
weasyprint
django-object-actions
django-rest-framework
celery
flower
django-celery-beat
django-ckeditor
django-3-jet
django-baton
channels